Native AMD64 build (Release|x64) alongside Win32; ASLR, DEP, high-entropy ASLR enabled; no VA ceiling
DC_MESSAGE_WIRE / ONLINEDATA_WIRE — fixed-width IPC structs replacing pointer-sized fields; qwBytesTransfered widened to UINT64; static_assert layout pins
IOCP completion key (Threads.c) — DWORD* to ULONG_PTR*; all negative sentinel values normalised to (ULONG_PTR)-N
IoAtomicExchangeSocket — replaces InterlockedExchange / InterlockedExchangePointer on SOCKET (UINT_PTR, 8 bytes on x64)
Tcl 9.0 Tcl_Size API — Tcl_GetStringFromObj length param widened from int* to Tcl_Size*; DC_VFS.pBuffer corrected from PBYTE[1] to BYTE[1]
DataCopy security hardening — per-command dwContextAvailable bounds check; SafeStringInRegion(); PROCESS_ALL_ACCESS to least-privilege; 32-bit WoW64 sender rejection; USERFILE_Zero_Internal / GROUPFILE_Zero_Internal wired in
Stack-walker and debug (IoDebug.c) — #ifdef _M_X64 selects Rip/Rbp/Rsp and IMAGE_FILE_MACHINE_AMD64; Memory.c debug header layout fixed
Post-port truncation audit — Tcl_WaitObject pointer stored as Tcl_WideInt (both paths); io var get/unset (LONG) cast removed from struct-offset arithmetic; spinlock InterlockedExchange((LPLONG)&lpData) to InterlockedExchangePointer
MessageWindow compatibility: 32bit apps cannot call 64bit ioFTPD MessageWindow and vice-versa

+++ src/Array.c
@@ -161,7 +161,7 @@ LPVOID QuickDelete(LPVOID *List, INT Items, LPVOID Item, QUICKCOMPAREPROC Comp, 
 
 		if (! Check || ! Check(Return))
 		{
-			iSize	= &List[Items] - &Result[1];
+			iSize	= (INT)(&List[Items] - &Result[1]);
 			MoveMemory(&Result[0], &Result[1], iSize * sizeof(LPVOID));
 		}
 		return Return;
@@ -183,7 +183,7 @@ LPVOID QuickDeleteIndex(LPVOID *List, INT Items, INT dwPos)
 	Result = &List[dwPos-1];
 	Return = Result[0];
 
-	iSize	= &List[Items] - &Result[1];
+	iSize	= (int)(&List[Items] - &Result[1]);
 	MoveMemory(&Result[0], &Result[1], iSize * sizeof(LPVOID));
 	return Return;
 }
